+++ core/src/test/scala/better/files/benchmarks/Benchmark.scala
@@ -1,8 +1,8 @@
 package better.files.benchmarks
 
-import org.scalatest.FunSuite
+import org.scalatest.funsuite.AnyFunSuite
 
-trait Benchmark extends FunSuite {
+trait Benchmark extends AnyFunSuite {
   def profile[A](f: => A): (A, Long) = {
     val t = System.nanoTime()
     (f, ((System.nanoTime() - t) / 1e6).toLong)